More closing channels
- closing-channels-non-atomic example is non-atomic, and channel could
be closed or emptied by another thread, raising an exception.

- closing-channels-nil uses just one channel and channel.receive? and
is atomic, but does not allow nil values, since its used to signal
done. Should be fastest. Consumer moved to main thread to block.

- closing-channels-exceptions  uses just one channel and raises when
channel is closed. Allows for nil values, but may be slower. Consumer
moved to main thread to block.